test: cleanup and prepare oprf-test crate
